+++ plugins/newspack-network/CHANGELOG.md
@@ -1,3 +1,10 @@
+## newspack-network [2.20.2](https://github.com/Automattic/newspack-workspace/compare/newspack-network@2.20.1...newspack-network@2.20.2) (2026-06-01)
+
+
+### Bug Fixes
+
+* **network:** harden the user-sync receive side and sign the Hub pull response ([#319](https://github.com/Automattic/newspack-workspace/issues/319)) ([0997396](https://github.com/Automattic/newspack-workspace/commit/0997396c7cdd9a69e62d9dfd29066dfb35a6fe8e))
+
 ## [2.20.1](https://github.com/Automattic/newspack-network/compare/v2.20.0...v2.20.1) (2026-05-18)